Warn against illegal SC.TextFieldView configurations
It's not currently possible to implement a text field that hides its value
for a multiline input. Likewise, a text field with a type other than 'text'
isn't possible for multiline inputs or hidden-value inputs. This commit adds
documentation to the comments of the `isPassword` and `type` properties warning
of these limitations, and also a developer warning when such a situation is
encountered.